requirements.txt
fastapi
uvicorn
requests
google-generativeai
supabase
python-dotenv
pydantic